Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float b;
-    scanf("%f",&b);
-    printf("You entered: %.4f",b);
+    double b;
+    scanf("%lf",&b);
+    printf("You entered: %lf",b);
     return 0;
 }
